requirements.txt
Flask>=2.3.0
Flask-Session>=0.5.0
psycopg2
validate_email>=1.3
qrcode[pil]>=7.3
Flask-Mail>=0.9.1
python-dotenv>=1.0.0
Werkzeug>=2.3.0